backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astroid==4.0.2
bcrypt==5.0.0
click==8.3.1
coverage==7.12.0
dill==0.4.0
dotenv==0.9.9
fastapi==0.121.2
flake8==7.3.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
isort==7.0.0
mccabe==0.7.0
mypy==1.18.2
mypy-extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
py-spy==0.4.1
pycodestyle==2.14.0
pydantic==2.12.4
pydantic-core==2.41.5
pyflakes==3.4.0
pygments==2.19.2
pyinstrument==5.1.1
pyjwt==2.10.1
pylint==4.0.3
pytest==9.0.1
python-dotenv==1.2.1
python-multipart==0.0.20
sniffio==1.3.1
sqlalchemy==2.0.44
starlette==0.49.3
tomlkit==0.13.3
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.38.0
